% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deNovoStructure.R
\name{deNovoStructureParam}
\alias{deNovoStructureParam}
\title{de novo structure annotation}
\usage{
deNovoStructureParam(numberOfCandidateToPredict = 10)
}
\arguments{
\item{numberOfCandidateToPredict}{\code{numeric}, number of structure candidates
to be predicted by MsNovelist. Max Value \code{128}. Actual number of
returned candidate might be lower du to duplicates being created.
Default is \code{10}}
}
\value{
An object of class \code{deNovoStructureParam}.
}
\description{
This function to set up the parameter for de novo structure annotation using
the \emph{MSNovelist} tool.

\emph{MSNovelist} generates molecular structures de novo from MS/MS data -
without relying on any database. This makes it particularly useful for
analyzing poorly represented analyte classes and novel compounds, where
traditional database searches may fall short. However, it is not intended
to replace database searches altogether, as structural elucidation of small
molecules from MS/MS data remains a challenging task, and identifying a
structure without database candidates is even more difficult.
}
\note{
For more information, see the Sirius
\href{https://v6.docs.sirius-ms.io/methods-background}{documentation}.
}
\examples{
# Example of setting up the parameters for de novo structure annotation
param <- deNovoStructureParam(numberOfCandidateToPredict = 10)

}
\references{
reference
}
