% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{summary}
\title{Fetch Results from Sirius}
\usage{
summary(
  sirius,
  result.type = c("formulaId", "structure", "deNovo", "spectralDbMatch")
)

results(
  sirius,
  features = character(),
  result.type = c("formulaId", "structureDb", "compoundClass", "deNovo",
    "spectralDbMatch", "fragTree"),
  return.type = c("list", "data.frame"),
  topFormula = 5,
  topStructure = 5,
  topSpectralMatches = 5
)
}
\arguments{
\item{sirius}{A \code{Sirius} object.}

\item{result.type}{\code{character(1)} specifying the type of results to fetch.
Options are "formulaId", "structureDb", "compoundClass", "deNovo",
"spectralDbMatch", and "fragTree". Defaults to "formulaId".}

\item{features}{\code{character()} vector specifying feature IDs to retrieve
results for. Defaults to all available features.}

\item{return.type}{\code{character(1)} specifying return format. Either
"data.frame" (default) or "list".}

\item{topFormula}{\code{numeric(1)} Maximum number of formula candidates to
retrieve per feature. Defaults to 5.}

\item{topStructure}{\code{numeric(1)} Maximum number of structure candidates per
formula. Defaults to 5.}

\item{topSpectralMatches}{\code{numeric(1)} Maximum number of spectral matches
per feature. Defaults to 5.}
}
\value{
A data.frame or list of results, depending on \code{return.type}.

a data.frame with the summary of the results.
Important column is the \emph{ApproximateConfidence} column, which give a
score of how possible all the identifications for this feature are.
The \emph{exactConfidence} column is a score of how possible the top
identification is.
}
\description{
Functions to retrieve results from the \emph{Sirius} project, allowing
customization of the result type, feature selection, and return format.

Get best match results for a Sirius project. Choose which type
of results to fetch with the \emph{result.type} parameter.
"structure" and "deNovo" will both also give information on
the formula.
}
