% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{config}
\title{Run job on Sirius.}
\usage{
run(
  sirius,
  compoundsIds = character(),
  alignedFeaturesIds = featuresId(sirius),
  fallbackAdducts = c("[M + H]+", "[M - H]-", "[M + Na]+", "[M + K]+"),
  enforceAdducts = character(),
  detectableAdducts = c("[M + H3N + H]+", "[M - H4O2 + H]+", "[M - H2O - H]-",
    "[M - H3N - H]-", "[M + Cl]-", "[2M + K]+", "[M + K]+", "[2M + Cl]-",
    "[M + C2H4O2 - H]-", "[M + H]+", "[2M + H]+", "[M - CH3 - H]-", "[M - H]-",
    "[M + Na]+", "[M - H2O + H]+"),
  spectraSearchParams = NA,
  formulaIdParams = NA,
  zodiacParams = NA,
  predictParams = NA,
  structureDbSearchParams = NA,
  msNovelistParams = NA,
  recompute = FALSE,
  configFile = character(),
  wait = TRUE
)

config(
  compoundsIds = character(),
  alignedFeaturesIds = character(),
  fallbackAdducts = c("[M + H]+", "[M - H]-", "[M + Na]+", "[M + K]+"),
  enforceAdducts = character(),
  detectableAdducts = c("[M + H3N + H]+", "[M - H4O2 + H]+", "[M - H2O - H]-",
    "[M - H3N - H]-", "[M + Cl]-", "[2M + K]+", "[M + K]+", "[2M + Cl]-",
    "[M + C2H4O2 - H]-", "[M + H]+", "[2M + H]+", "[M - CH3 - H]-", "[M - H]-",
    "[M + Na]+", "[M - H2O + H]+"),
  formulaIdParams = formulaIdParam(),
  zodiacParams = NA,
  predictParams = NA,
  structureDbSearchParams = NA,
  msNovelistParams = NA,
  spectraSearchParams = NA,
  recompute = FALSE
)
}
\arguments{
\item{sirius}{\code{Sirius} object, the connection to the Sirius server.}

\item{compoundsIds}{\code{character} vector, the ids of the compounds to process.}

\item{alignedFeaturesIds}{\code{character} vector, the ids of the aligned features
to process. By default, computes all.}

\item{fallbackAdducts}{\code{character} vector, fallback adducts are considered
if the auto detection did not find any indication for an ion mode.}

\item{enforceAdducts}{\code{character} vector, the adducts to enforce. They are
always considered.}

\item{detectableAdducts}{\code{character} vector, detectable adducts are only
considered if there is an indication in the MS1 scan (e.g. correct
mass delta).}

\item{spectraSearchParams}{object of class \code{spectraMatchingParam}, containing
the parameters for the spectra matching.}

\item{formulaIdParams}{object of class \code{formulaIdParam}, containing the
parameters for the molecular formula identification.}

\item{zodiacParams}{object of class \code{zodiacParam}, containing the parameters
for the Zodiac re-ranking.}

\item{predictParams}{object of class \code{predictParam}, containing the parameters for
the molecular fingerprint prediction and compound classification.}

\item{structureDbSearchParams}{object of class \code{structureDbParam}, containing
the parameters for the structure annotation.}

\item{msNovelistParams}{object of class \code{deNovoStructureParam}, containing the
parameters for the de novo structure generation.}

\item{recompute}{\code{logical}, whether to recompute the job , only necessary if
the configfile comes from a saved file. Default is FALSE.}

\item{configFile}{\code{character}, the path to the configuration file to use for
the job submission, optional.}

\item{wait}{\code{logical}, whether to wait for the job to finish. Default is TRUE}
}
\value{
The job ID of the submitted job, it can be inputted in the
\code{jobInfo()} function to retrieve the job information. To retrieve results
see \code{\link{results}} documentation.
}
\description{
This function configures the job submission to Sirius.
It creates an object of class \code{config} that can be used to submit a job
Sirius, it can also be saved and reused later on through the
\code{saveJobConfig()}function. For example on how to use, see the vignette.

Depending on what task you want to perform, you can specify the following
parameters:
\itemize{
\item \code{\link{spectraMatchingParam}}: Allows to perform matching betweeen spectra input
to spectral libraries.
\item \code{\link{formulaIdParam}}: Allows to generate molecular formula candidates for
each features.
\item \code{\link{zodiacParam}}: Allows to perform re-ranking of formula candidates
using \emph{Zodiac}. It is advised to only perform it if De Novo structure
annotation is run later.
\item \code{\link{predictParam}}: Allows to perform molecular fingeprint prediction using
\emph{CSI:FingerID} and compound classification using \emph{CANOPUS}.
\item \code{\link{structureDbSearchParam}}: Allows to perform structure annotation based
on the fingerprint identifications.
\item \code{\link{deNovoStructureParam}}: Allows to perform de novo structure generation using
the \emph{MSNovelist} tool.
}
}
\section{Only formula identification}{


If you only want to perform formula identification, you can by only inputing
the \code{formulaIdParam} object. In combination, you can also input the
\code{spectraMatchingParam} object to perform spectral matching and subsequently
compare the results.
}

\section{Structure annotation}{


To performe structure annotation, you need input the \code{formulaIdParam} object,
as well as the \code{predictParam} object to perform molecular fingerprint prediction
and compound classification. These results will then subsequently be used to
perform structure annotation using the \code{structureDbSearchParam} object.
}

\section{De Novo structure annotation}{


To perform de novo structure annotation, you need to input the \code{formulaIdParam}
object, it is also advised in this case to perform re-ranking using the
\code{zodiacParam} object. The molecular fingerprint prediction and compound
classification can be performed using the \code{predictParam} object.
The \code{deNovoStructureParam} object is then used to perform the de novo structure
annotation.
}

