% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siriusDBs.R
\name{siriusDBs}
\alias{siriusDBs}
\alias{listDBs}
\alias{infoDb}
\alias{removeDb}
\alias{createDB}
\title{Functions for Handling Sirius Databases}
\usage{
listDBs(sirius)

infoDb(sirius, databaseId = character())

removeDb(sirius, databaseId = character())

createDB(
  sirius,
  databaseId = character(),
  files = character(),
  location = getwd()
)
}
\arguments{
\item{sirius}{A \code{Sirius} object representing the Sirius connection.}

\item{databaseId}{A \code{character(1)} string specifying the new database ID.}

\item{files}{A \code{character} vector of file paths to add to the database.}

\item{location}{A \code{character(1)} string specifying the location for the
database. Defaults to the current working directory.}
}
\value{
A \code{data.frame} containing details of searchable databases.

A \code{list} containing details of the specified database.

A \code{list} containing details of the created database.
}
\description{
This set of functions provides tools for managing databases in Sirius. These
include listing available databases, retrieving database information,
deleting databases, and creating new databases from files.

List the databases that are searchable by Sirius for spectral matching.

Retrieve detailed information about a specific database.

Delete a database from Sirius.

Create a new database in Sirius from specified files.
}
