% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraMatching.R
\name{spectraMatchingParam}
\alias{spectraMatchingParam}
\title{Spectra database matching}
\usage{
spectraMatchingParam(
  spectraSearchDBs = c("BIO", "massbank"),
  peakDeviationPpm = 10,
  precursorDeviationPpm = 10,
  scoring = c("MODIFIED_COSINE", "INTENSITY", "GAUSSIAN")
)
}
\arguments{
\item{spectraSearchDBs}{\code{character} vector of the databases to search. The
default is \code{c("BIO", "massbank")}. Other values can be found by
running \code{listDatabases()}.}

\item{peakDeviationPpm}{\code{numeric} Maximum allowed mass deviation in ppm for
matching peaks.}

\item{precursorDeviationPpm}{\code{numeric} Maximum allowed mass deviation in ppm
for matching the precursor. If not specified, the same value as for
the peaks is used.}

\item{scoring}{\code{character} The scoring function to use. Possible values are
\code{MODIFIED_COSINE}, \code{GAUSSIAN}, \code{INTENSITY}. The default is
\code{MODIFIED_COSINE}.}
}
\value{
An object of class \code{spectraMatchingParam}
}
\description{
This function is to set up the parameter for matching to spectra
databases. This needs to be run first. Spectral library matching is
performed using the cosine score with squared peak intensities, ignoring the
precursor peak.

Note that spectral library matches are added as annotations to CSI:FingerID
results and do not influence the ranking of structure candidates.
}
\note{
For more information, see the Sirius
\href{https://v6.docs.sirius-ms.io/methods-background}{documentation}.
}
\examples{
# Example of setting up the parameters for spectra matching
param <- spectraMatchingParam(spectraSearchDBs = c("BIO", "massbank"),
                              peakDeviationPpm = 10,
                              precursorDeviationPpm = 10,
                              scoring = "MODIFIED_COSINE")

}
