% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureDbSearch.R
\name{structureDbSearchParam}
\alias{structureDbSearchParam}
\title{Structure Database Search}
\usage{
structureDbSearchParam(
  structureSearchDbs = c("BIO", "massbank"),
  tagStructuresWithLipidClass = TRUE,
  expansiveSearchConfidenceMode = c("APPROXIMATE", "EXACT", "OFF")
)
}
\arguments{
\item{structureSearchDbs}{\code{character} vector, specifying the id of the
databases to search for structures. Default is \code{c("BIO", "massbank")}.
If expansive search is enabled this database selection will be
expanded to PubChem if not high confidence hit was found in the
selected databases. The available databases can be found by running
\code{listDatabases()}.}

\item{tagStructuresWithLipidClass}{\code{logical}, specifying whether to tag
structures with lipid class estimated by El Gordo. The lipid class
will only be available if El Gordo predicts that the MS/MS  is a
lipid spectrum. If this parameter is set to \code{FALSE} El Gordo will
still be executed and e.g. improve the fragmentation tree, but the
matching structure candidates will not be tagged if they match lipid
class.}

\item{expansiveSearchConfidenceMode}{\code{character(1)} , specifying the
confidence mode for expansive search. Possible values are
\code{"APPROXIMATE"},\code{"EXACT"}, \code{"OFF"}. Defaults to \code{"APPROXIMATE"}.}
}
\value{
An object of class \code{structureDbSearchParam}.
}
\description{
This function creates an object of class \code{structureDbSearchParam} that can
be used to configure the structure database search in Sirius. The object can
be passed to the \code{runSirius} function to perform the structure database
search.

By default, \emph{SIRIUS} searches for molecular structures in a biomolecule
structure database. It can also search in the (extremely large) PubChem
database or in custom “suspect databases” provided by the user.
}
\note{
For more information, see the Sirius
\href{https://v6.docs.sirius-ms.io/methods-background}{documentation}.
}
\examples{
# Example of setting up the parameters for structure database search
param <- structureDbSearchParam(
                 structureSearchDbs = c("BIO", "massbank"),
                 tagStructuresWithLipidClass = TRUE,
                 expansiveSearchConfidenceMode = "APPROXIMATE")


}
\references{
reference
}
