% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zodiacParam.R
\name{zodiacParam}
\alias{zodiacParam}
\title{Configuration fr re-ranking of Molecular Formula Annotation}
\usage{
zodiacParam(
  consideredCandidatesAt300Mz = 10,
  consideredCandidatesAt800Mz = 50,
  runInTwoSteps = TRUE,
  edgeFilterThreshold = TRUE,
  thresholdFilter = 0.95,
  minLocalCandidates = 1,
  minLocalConnections = 10,
  gibbsSamplerParameters = TRUE,
  iterations = 20000,
  burnInPeriod = 2000,
  numberOfMarkovChains = 10
)
}
\arguments{
\item{consideredCandidatesAt300Mz}{An \code{integer(1)} specifying the maximum
number of candidate molecular formulas considered by ZODIAC for
compounds below 300 m/z. Default is \code{10}.}

\item{consideredCandidatesAt800Mz}{An \code{integer(1)} specifying the maximum
number of candidate molecular formulas considered by ZODIAC for \
compounds above 800 m/z. Default is \code{50}.}

\item{runInTwoSteps}{\code{logical} specifying whether ZODIAC uses a
2-step approach. First running 'good quality compounds' only, and
afterwards including the remaining compounds. Default is \code{TRUE}.}

\item{edgeFilterThreshold}{\code{logical} value specifying whether ZODIAC uses
an edge filter threshold. Default is \code{TRUE}.}

\item{thresholdFilter}{\code{numeric(1)} value specifying the threshold filter.
Default is \code{0.95}.}

\item{minLocalCandidates}{\code{numeric(1)} value specifying the minimum number of
local candidates. Default is \code{1}.}

\item{minLocalConnections}{\code{numeric(1)} value specifying the minimum number of
local connections. Default is \code{10}.}

\item{gibbsSamplerParameters}{\code{logical} value specifying whether ZODIAC
uses Gibbs sampler parameters. Default is \code{TRUE}.}

\item{iterations}{\code{numeric(1)} specifying the number of iterations. Default is
\code{20000}.}

\item{burnInPeriod}{\code{numeric(1)} specifying the burn-in period. Default is
\code{2000}.}

\item{numberOfMarkovChains}{\code{numeric(1)} specifying the number of Markov
chains. Default is \code{10}.}
}
\value{
An object of class \code{zodiacParam}.
}
\description{
This function configures the parameters for the re-ranking of the previously
computed molecular formula annotation in Sirius. This step is quite
computationally and memory demanding, it is advised to perform it only if
de novo structure annotation is used later on

ZODIAC uses the top X molecular formula candidates for each molecule from
SIRIUS to build a similarity network, and uses Bayesian statistics to
re-rank those candidates.
}
\note{
For more information, see the Sirius
\href{https://v6.docs.sirius-ms.io/methods-background}{documentation}.
}
\examples{
# Example of setting up the parameter for Zodiac re-ranking
param <- zodiacParam(consideredCandidatesAt300Mz = 10,
                     consideredCandidatesAt800Mz = 50,
                     runInTwoSteps = TRUE,
                     edgeFilterThreshold = TRUE,
                     thresholdFilter = 0.95,
                     minLocalCandidates = 1,
                     minLocalConnections = 10,
                     gibbsSamplerParameters = TRUE,
                     iterations = 20000,
                     burnInPeriod = 2000,
                     numberOfMarkovChains = 10)

}
\references{
reference
}
