% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import Data into Sirius}
\usage{
import(sirius, ms1Spectra, ms2Spectra)
}
\arguments{
\item{sirius}{\code{Sirius}, the connection to the Sirius instance with a
loaded project}

\item{ms1Spectra}{\code{Spectra}, the MS1 spectra to import}

\item{ms2Spectra}{\code{Spectra}, the MS2 spectra to import}
}
\description{
Import Data into Sirius
}
\note{
Will remove the previously imported feature.
}
