% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictParam.R
\name{predictParam}
\alias{predictParam}
\title{Predicting FingerPrint and compounds Identifications}
\usage{
predictParam(useScoreThreshold = TRUE, alwaysPredictHighRefMatches = FALSE)
}
\arguments{
\item{useScoreThreshold}{\code{logical} whether to use a soft threshold to be
applied to only compute FingerPrints for promising formula candidates.
Enabling is highly recommended. Default is \code{TRUE}.}

\item{alwaysPredictHighRefMatches}{\code{logical} wether to predict
FingerPrint/Classes&Structure for formulas candidates with reference
spectrum similarity > \code{Sirius.minReferenceMatchScoreToInject} no
matter which score threshold rules apply. Default is \code{FALSE}.}
}
\description{
This function creates an object of class \code{predictParam} that can be used to
predict molecular fingerprints and compound identifications using
\emph{CSI:FingerID} and \emph{CANOPUS}.

CSI:FIngerID identifies the structure of a molecule by predicting its
molecular fingerprint and using this fingerprint to search in a molecular
structure database.

CANOPUS (Dührkop et al.) predicts the presense/absense of more than 2500
compound classes. CANOPUS predicts these classes based solely on MS/MS data
and without requiring database information. This means it can identify a
class even if no molecular structure of that class exists in the molecular
structure database.
}
\note{
For more information, see the Sirius \href{https://v6.docs.sirius-ms.io/methods-background}{documentation}.
}
\examples{
# Example of setting up the parameters for the prediction of molecular
# fingerprints and compound class
param <- predictParam()

}
\references{
reference
}
