% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{summary}
\title{Get results from Sirius}
\usage{
summary(sirius, deNovo = FALSE)

results(
  sirius,
  features = character(),
  result.type = c("formulaId", "structureDb", "compoundClass", "deNovo"),
  return.type = c("list", "data.frame"),
  topFormula = 5,
  topStructure = 5
)
}
\arguments{
\item{sirius}{a \code{Sirius} object}

\item{deNovo}{\code{logical(1)} specifying if the summary results should be fetched
for the de novo candidates. Defaults to \code{FALSE}.}

\item{features}{\code{character()} vector specifying the feature IDs to retrieve
results for. If not provided, all available features will be used.}

\item{result.type}{\code{character(1)} specifying the type of results to retrieve.
Possible values are "formulaId", "structureDb", "compoundClass",
and "deNovo". Defaults to "formulaId". All the possible values will
return  formulaId results anyway.}

\item{return.type}{\code{character(1)} specifying the return type. Either
\code{"data.frame"}, which will return a single \code{data.frame} with all
results. Or \code{"list"}, which will return a \code{list} of \code{data.frame}
where each element corresponds to the results for a single feature.
Defaults to \code{"data.frame"}.}

\item{topFormula}{\code{numeric(1)} specifying the maximum number of formula
candidates to retrieve for each feature.}

\item{topStructure}{\code{numeric(1)} specifying the maximum number of structure
candidates to retrieve for each formula.}
}
\value{
a data.frame with the summary of the results.
Important column is the \emph{ApproximateConfidence} column, which give a
score of how possible all the identifications for this feature are.
The \emph{exactConfidence} column is a score of how possible the top
identification is.
}
\description{
The function below are used to fetch results from the \emph{Sirius} project.
}
