% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Utils}
\alias{logIn}
\alias{checkConnection}
\alias{shutdown}
\alias{openWebApi}
\alias{openGUI}
\alias{closeGUI}
\alias{projectInfo}
\alias{listOpenProjects}
\alias{openProject}
\alias{featuresId}
\alias{featuresInfo}
\alias{deleteFeatures}
\alias{mapFeatures}
\alias{saveConfig}
\alias{jobInfo}
\alias{deleteJob}
\title{Utility function for RuSirius}
\usage{
logIn(sirius, username, password)

checkConnection(sirius)

shutdown(sirius, closeProject = TRUE)

openWebApi(sirius)

openGUI(sirius)

closeGUI(sirius)

projectInfo(sirius, infoType = c("compatibilityInfo", "sizeInformation"))

listOpenProjects(sirius)

openProject(sirius, projectId, path = character())

featuresId(sirius, type = c("sirius", "xcms"))

featuresInfo(sirius)

deleteFeatures(sirius, featureId = featuresId(sirius))

mapFeatures(sirius)

saveConfig(sirius, config, name)

jobInfo(sirius, jobId = character())

deleteJob(sirius, jobId = character(), all = FALSE)
}
\arguments{
\item{sirius}{a \code{Sirius} object}

\item{username}{\code{character(1)}, the username to use for the connection}

\item{password}{\code{character(1)}, the password to use for the connection}

\item{closeProject}{\code{logical}, whether to close the project before
shutting down Sirius. Default is \code{TRUE}.}

\item{infoType}{\code{character} vector of length 1 or 2, specifying the type of
information to retrieve. Possible values are \code{"compatibilityInfo"} and
\code{"sizeInformation"}.}

\item{projectId}{\code{character(1)} specifying the id of the project to open.
can be an already existing project or a new one.}

\item{path}{\code{character(1)} path where to find the existing project or where
to create a new one.By default, the porject will be opened in the
current \code{"."} directory.}

\item{type}{\code{character} vector of length 1, specifying the type of features
ID to list. Possible values are \code{"sirius"} and \code{"xcms"}. Default
is \code{"sirius"}.}

\item{featureId}{\code{character()} vector specifying the id of the feature to
remove. By default remove all.}

\item{config}{a \code{configSirius} object}

\item{name}{\code{character} vector of length 1, specifying the name of the
configuration to load.}

\item{jobId}{\code{character(1)} specifying the id of the job to delete.}

\item{all}{\code{logical}, whether to delete all jobs. Default is \code{FALSE}.}
}
\value{
A \code{message} whether the user is successfully logged in or not.

\code{logical}, \code{TRUE} if the connection is valid, \code{FALSE} otherwise.

\code{message} whether the project is closed or not.

nothing, will open the web API in the default browser.

\code{TRUE} if the GUI opens correctly.

\code{TRUE} if the GUI closes correctly.

a \code{list} with the information requested.

a \code{character} vector with the open projects.

a \code{Sirius} object with the project opened.

a \code{character} vector with the features ID.

a \code{data.frame} with the features information.

A \code{Sirius} object with the features removed.

a \code{data.frame} with the features mapping.

nothing, will save the configuration locally.

a \code{character} vector with the job information.

nothing, will delete the job from Sirius.
}
\description{
returns \code{TRUE} if the connection to the Sirius is valid,
\code{FALSE} otherwise.
}
